[staking]: Add getProposerValId() getter
After MONAD_FIVE, the reward syscall sets a storage slot with the
validator ID of the most recent proposer. This slot has the lifetime of
the block.

The new slot is used by the getProposerValId() getter. This is primarily
to support LST validator attribution. Without this getter, they need to
ask validators to hack their val id into the beneficiary field

The new getter is charged as a single warm sload. This is because the
reward function comes first and writes to the slots, so all readers will
read a cached value from BlockState, so disk is never accessed.